CHG : 명시적인 형태로 클래스 이름 변경
